Fixed share count issue on 32-bit builds
Also added step_factor option to manage GPU load.
Minor tweaking on OpenCL code.

+++ xptMiner/global.h
@@ -194,7 +194,7 @@ void xptMiner_submitShare(minerPrimecoinBlock_t* block);
 void xptMiner_submitShare(minerMetiscoinBlock_t* block);
 
 // stats
-extern volatile uint64 totalCollisionCount;
+extern volatile uint32 totalCollisionCount;
 extern volatile uint32 totalShareCount;
 extern volatile uint32 invalidShareCount;
 extern volatile uint32 monitorCurrentBlockHeight;
